feat: normalize reviewpad file
